Refactor DdAndroidGradlePluginTest into focused test files
Split the monolithic test file (1493 lines) into smaller, focused test files organized by method under test:

- DdAndroidGradlePluginTestBase.kt: Shared setup and helper methods
- DdAndroidGradlePluginResolveApiKeyTest.kt: API key resolution tests (13 tests)
- DdAndroidGradlePluginResolveExtensionConfigurationTest.kt: Extension config tests (14 tests)
- DdAndroidGradlePluginConfigureVariantForUploadTest.kt: Upload task configuration tests (11 tests)
- DdAndroidGradlePluginConfigureVariantForSdkCheckTest.kt: SDK check tests (4 tests)
